@@ -9,9 +9,8 @@ namespace Millistream.Streaming
     /// Represents a managed message handle (mdf_message_t) that can contain several messages for efficiency.
     /// </summary>
     /// <remarks>Handles are not thread-safe. If multiple threads will share access to a single handle, the accesses has to be serialized using a mutex or other forms of locking mechanisms. The API as such is thread-safe so multiple threads can have local handles without the need for locks.</remarks>
-    public sealed class Message : IMessage, IDisposable
+    public unsafe sealed class Message : IMessage, IDisposable
     {
-        private readonly INativeImplementation _nativeImplementation;
         private CompressionLevel _compressionLevel = CompressionLevel.Z_BEST_SPEED;
         private bool _utf8Validation = true;
         private bool _isDisposed;
@@ -21,13 +20,7 @@ namespace Millistream.Streaming
         /// </summary>
         /// <exception cref="DllNotFoundException">The native dependency is missing.</exception>
         /// <remarks>The corresponding native function is mdf_message_create.</remarks>
-        public Message() : this(NativeImplementation.Instance) { }
-
-        internal Message(INativeImplementation nativeImplementation)
-        {
-            _nativeImplementation = nativeImplementation ?? throw new ArgumentNullException(nameof(nativeImplementation));
-            Handle = _nativeImplementation.mdf_message_create();
-        }
+        public Message() => Handle = NativeImplementation.mdf_message_create();
 
         ~Message() => Dispose();
 
@@ -46,7 +39,7 @@ namespace Millistream.Streaming
             set
             {
                 ThrowIfDisposed();
-                if (_nativeImplementation.mdf_message_set_compression_level(Handle, (int)value) == 1)
+                if (NativeImplementation.mdf_message_set_compression_level(Handle, (int)value) == 1)
                     _compressionLevel = value;
             }
         }
@@ -60,7 +53,7 @@ namespace Millistream.Streaming
             get
             {
                 ThrowIfDisposed();
-                return _nativeImplementation.mdf_message_get_num(Handle);
+                return NativeImplementation.mdf_message_get_num(Handle);
             }
         }
 
@@ -73,7 +66,7 @@ namespace Millistream.Streaming
             get
             {
                 ThrowIfDisposed();
-                return _nativeImplementation.mdf_message_get_num_active(Handle);
+                return NativeImplementation.mdf_message_get_num_active(Handle);
             }
         }
 
@@ -91,7 +84,7 @@ namespace Millistream.Streaming
             set
             {
                 ThrowIfDisposed();
-                if (_nativeImplementation.mdf_message_set_utf8_validation(Handle, value ? 1 : 0) == 1)
+                if (NativeImplementation.mdf_message_set_utf8_validation(Handle, value ? 1 : 0) == 1)
                     _utf8Validation = value;
             }
         }
@@ -109,7 +102,7 @@ namespace Millistream.Streaming
         public bool Add(ulong insref, int mref)
         {
             ThrowIfDisposed();
-            return _nativeImplementation.mdf_message_add(Handle, insref, mref) == 1;
+            return NativeImplementation.mdf_message_add(Handle, insref, mref) == 1;
         }
 
         /// <summary>
@@ -144,7 +137,7 @@ namespace Millistream.Streaming
                 byte* bytes = stackalloc byte[value.Length + 1];
                 if (!TryGetAsciiBytes(value, bytes))
                     return false;
-                return _nativeImplementation.mdf_message_add_numeric(Handle, tag, (IntPtr)bytes) == 1;
+                return NativeImplementation.mdf_message_add_numeric(Handle, tag, (IntPtr)bytes) == 1;
             }
         }
 
@@ -176,7 +169,7 @@ namespace Millistream.Streaming
             if (decimals < 0 || decimals > 19)
                 throw new ArgumentException($"{nameof(decimals)} cannot be smaller than 0 or greater than 19.", nameof(decimals));
             ThrowIfDisposed();
-            return _nativeImplementation.mdf_message_add_int(Handle, tag, value, decimals) == 1;
+            return NativeImplementation.mdf_message_add_int(Handle, tag, value, decimals) == 1;
         }
 
         /// <summary>
@@ -209,7 +202,7 @@ namespace Millistream.Streaming
             if (decimals < 0 || decimals > 19)
                 throw new ArgumentException($"{nameof(decimals)} cannot be smaller than 0 or greater than 19.", nameof(decimals));
             ThrowIfDisposed();
-            return _nativeImplementation.mdf_message_add_uint(Handle, tag, value, decimals) == 1;
+            return NativeImplementation.mdf_message_add_uint(Handle, tag, value, decimals) == 1;
         }
 
         /// <summary>
@@ -250,7 +243,7 @@ namespace Millistream.Streaming
                     byte* b = stackalloc byte[length + 1];
                     int bytesWritten = Encoding.UTF8.GetBytes(c, value.Length, b, length);
                     b[bytesWritten] = 0;
-                    return _nativeImplementation.mdf_message_add_string(Handle, tag, (IntPtr)b) == 1;
+                    return NativeImplementation.mdf_message_add_string(Handle, tag, (IntPtr)b) == 1;
                 }
             }
         }
@@ -283,7 +276,7 @@ namespace Millistream.Streaming
                     int byteCount = Encoding.UTF8.GetByteCount(c, length);
                     byte* b = stackalloc byte[byteCount + 1];
                     Encoding.UTF8.GetBytes(c, length, b, byteCount);
-                    return _nativeImplementation.mdf_message_add_string2(Handle, tag, (IntPtr)b, byteCount) == 1;
+                    return NativeImplementation.mdf_message_add_string2(Handle, tag, (IntPtr)b, byteCount) == 1;
                 }
             }
         }
@@ -335,7 +328,7 @@ namespace Millistream.Streaming
                 byte* bytes = stackalloc byte[value.Length + 1];
                 if (!TryGetAsciiBytes(value, bytes))
                     return false;
-                return _nativeImplementation.mdf_message_add_date(Handle, tag, (IntPtr)bytes) == 1;
+                return NativeImplementation.mdf_message_add_date(Handle, tag, (IntPtr)bytes) == 1;
             }
         }
 
@@ -365,7 +358,7 @@ namespace Millistream.Streaming
         public bool AddDate(uint tag, int year, int month, int day)
         {
             ThrowIfDisposed();
-            return _nativeImplementation.mdf_message_add_date2(Handle, tag, year, month, day) == 1;
+            return NativeImplementation.mdf_message_add_date2(Handle, tag, year, month, day) == 1;
         }
 
         /// <summary>
@@ -403,7 +396,7 @@ namespace Millistream.Streaming
                 byte* bytes = stackalloc byte[value.Length + 1];
                 if (!TryGetAsciiBytes(value, bytes))
                     return false;
-                return _nativeImplementation.mdf_message_add_time(Handle, tag, (IntPtr)bytes) == 1;
+                return NativeImplementation.mdf_message_add_time(Handle, tag, (IntPtr)bytes) == 1;
             }
         }
 
@@ -434,7 +427,7 @@ namespace Millistream.Streaming
         public bool AddTime2(uint tag, int hour, int minute, int second, int millisecond)
         {
             ThrowIfDisposed();
-            return _nativeImplementation.mdf_message_add_time2(Handle, tag, hour, minute, second, millisecond) == 1;
+            return NativeImplementation.mdf_message_add_time2(Handle, tag, hour, minute, second, millisecond) == 1;
         }
 
         /// <summary>
@@ -467,7 +460,7 @@ namespace Millistream.Streaming
         public bool AddTime3(uint tag, int hour, int minute, int second, int nanosecond)
         {
             ThrowIfDisposed();
-            return _nativeImplementation.mdf_message_add_time3(Handle, tag, hour, minute, second, nanosecond) == 1;
+            return NativeImplementation.mdf_message_add_time3(Handle, tag, hour, minute, second, nanosecond) == 1;
         }
 
         /// <summary>
@@ -510,7 +503,7 @@ namespace Millistream.Streaming
                 byte* bytes = stackalloc byte[value.Length + 1];
                 if (!TryGetAsciiBytes(value, bytes))
                     return false;
-                return _nativeImplementation.mdf_message_add_list(Handle, tag, (IntPtr)bytes) == 1;
+                return NativeImplementation.mdf_message_add_list(Handle, tag, (IntPtr)bytes) == 1;
             }
         }
 
@@ -538,7 +531,7 @@ namespace Millistream.Streaming
         public void Reset()
         {
             ThrowIfDisposed();
-            _nativeImplementation.mdf_message_reset(Handle);
+            NativeImplementation.mdf_message_reset(Handle);
         }
 
         /// <summary>
@@ -550,7 +543,7 @@ namespace Millistream.Streaming
         public bool Delete()
         {
             ThrowIfDisposed();
-            return _nativeImplementation.mdf_message_del(Handle) == 1;
+            return NativeImplementation.mdf_message_del(Handle) == 1;
         }
 
         /// <summary>
@@ -564,8 +557,13 @@ namespace Millistream.Streaming
         /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null" />.</exception>
         /// <exception cref="InvalidOperationException">The installed version of the native library doesn't include the mdf_message_move function.</exception>
         /// <remarks>The corresponding native function is mdf_message_move.</remarks>
-        public static bool Move(Message source, Message destination, ulong sourceInsref, ulong destinationInsRef) =>
-            Move(source, destination, sourceInsref, destinationInsRef, NativeImplementation.Instance);
+        public static bool Move(Message source, Message destination, ulong sourceInsref, ulong destinationInsRef)
+        {
+            if (source == null)
+                throw new ArgumentNullException(nameof(source));
+
+            return NativeImplementation.mdf_message_move(source.Handle, destination?.Handle ?? IntPtr.Zero, sourceInsref, destinationInsRef) == 1;
+        }
 
         /// <summary>
         /// Serializes the message chain in the message handle and produces a base64 encoded string to the address pointed to by <paramref name="result"/>. It's the responsibility of the caller to free the produced unmanaged string.
@@ -579,7 +577,7 @@ namespace Millistream.Streaming
         {
             ThrowIfDisposed();
             result = IntPtr.Zero;
-            return _nativeImplementation.mdf_message_serialize(Handle, ref result) == 1;
+            return NativeImplementation.mdf_message_serialize(Handle, ref result) == 1;
         }
 
         /// <summary>
@@ -602,7 +600,7 @@ namespace Millistream.Streaming
                 byte* bytes = stackalloc byte[data.Length + 1];
                 if (!TryGetAsciiBytes(data, bytes))
                     return false;
-                return _nativeImplementation.mdf_message_deserialize(Handle, (IntPtr)bytes) == 1;
+                return NativeImplementation.mdf_message_deserialize(Handle, (IntPtr)bytes) == 1;
             }
         }
 
@@ -617,7 +615,7 @@ namespace Millistream.Streaming
         public bool Deserialize(IntPtr data)
         {
             ThrowIfDisposed();
-            return _nativeImplementation.mdf_message_deserialize(Handle, data) == 1;
+            return NativeImplementation.mdf_message_deserialize(Handle, data) == 1;
         }
 
         /// <summary>
@@ -628,21 +626,10 @@ namespace Millistream.Streaming
         {
             if (!_isDisposed)
             {
-                _nativeImplementation?.mdf_message_destroy(Handle);
+                NativeImplementation.mdf_message_destroy(Handle);
                 _isDisposed = true;
                 GC.SuppressFinalize(this);
             }
-        }
-
-        internal static bool Move(Message source, Message destination, ulong insrefSource, ulong insrefDestination, INativeImplementation nativeImplementation)
-        {
-            if (source == null)
-                throw new ArgumentNullException(nameof(source));
-
-            if (nativeImplementation == null)
-                throw new ArgumentNullException(nameof(nativeImplementation));
-
-            return nativeImplementation.mdf_message_move(source.Handle, destination?.Handle ?? IntPtr.Zero, insrefSource, insrefDestination) == 1;
         }
 
         private void ThrowIfDisposed()
